# openapi.yaml
openapi: 3.0.3
info:
  title: CEC HTTP Bridge API
  description: |
    REST API for controlling HDMI-CEC devices (Rust implementation). Power,
    volume, source/HDMI switching, remote keys, raw CEC commands, MQTT
    settings, a dev console surface, and real-time events (SSE + WebSocket).
    Intended for home automation, media centers, and scripting.

    ## Envelope
    Every JSON response uses the envelope
    `{"status":"success"|"error","message":"...","data":...}`.
    `message` and `data` are omitted when empty/absent. Long-running
    operations return `success` with `data.accepted=true` (HTTP 202).

    ## Authentication
    Auth is **optional** and active only when the server was started with a
    token (`-token <secret>` / `auth_token` in `config.json`). When active,
    every `/api` route **except `GET /api/health`** (and `GET /metrics`)
    requires the token via `Authorization: Bearer <token>`,
    `X-Auth-Token: <token>`, `?key=<token>`, or the `capi_token` cookie set by
    the UI login page. Mutating cookie-authenticated requests additionally
    require same-origin. When no token is configured, no auth is required.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local (default)
  - url: http://{host}:8080
    description: Network
    variables:
      host:
        default: localhost
security:
  - BearerAuth: []
...
paths:
...
  /api/devices:
    get:
      tags: [Devices]
      summary: List CEC devices
      description: |
        Returns devices from the **bus steward cache** by default (immediate,
        no bus I/O). `?wait=N` (seconds, capped at 10) blocks up to N seconds
        for a fresh scan to land before serving the cache. `?live=1` /
        `?rescan=1` force a full bus reconcile before returning (may take
        tens of seconds on slow buses; the server responds 504 after ~50s).
        Each entry carries a `discovery` provenance (`active` = enumerated by
        libCEC, `polled` = confirmed by POLL probe, `observed` = ghost seen in
        passive bus traffic) plus `polled_at` / `first_seen_at` /
        `last_seen_at` timestamps and passive `observed_*` fields.
      operationId: getDevices
      parameters:
        - name: live
          in: query
          description: Force immediate full reconcile (same as rescan=1)
          required: false
          schema:
            type: string
            enum: ['1', 'true']
        - name: rescan
          in: query
          description: Force a full bus reconcile before returning
          required: false
          schema:
            type: string
            enum: ['1', 'true', 'false']
        - name: wait
          in: query
          description: >
            Max seconds to wait for a fresher scan before serving the cache.
            Values above 10 are clamped to 10. Default: immediate cache.
          required: false
          schema:
            type: number
            maximum: 10
      responses:
        '200':
          description: Devices retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                status: success
                message: Devices retrieved
                data:
                  - logical_address: 0
                    address_name: TV
                    physical_address: 0.0.0.0
                    device_type: TV
                    hdmi_port: 0
                    vendor_id: "0x0000F0"
                    vendor_name: Samsung
                    cec_version: "1.4"
                    power_status: On
                    osd_name: TV
                    menu_language: eng
                    is_active: true
                    is_active_source: false
                    discovery: active
                    polled_at: "2026-08-21T10:00:00.123456Z"
                    first_seen_at: "2026-08-21T09:31:02.000000Z"
                    last_seen_at: "2026-08-21T10:00:00.123456Z"
                    observed_power_status: "on"
                    observed_at: "2026-08-21T09:59:58.000000Z"
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: Bus steward queue full
        '504':
          description: Scan timed out (server limit ~50s; allow longer client timeouts)
  /api/devices/{address}:
    get:
      tags: [Devices]
      summary: Get device by logical address
      description: Get detailed information about a specific CEC device.
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Device info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────────── Bus ─────────────────────────────────
  /api/bus/state:
    get:
      tags: [Bus]
      summary: Bus steward snapshot
      description: |
        Cached CEC bus state including `devices` (same shape as
        `GET /api/devices`), `active_source`, `logical_addresses`, `stale`,
        `last_full_scan_at`, `scan_in_progress`, and optional `recent_frames`
        when `bus.frame_ring_size` is set in `config.json`. Served from the
        cached snapshot; topology refreshes happen asynchronously.
      operationId: getBusState
      responses:
        '200':
          description: Bus state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/bus/scan:
    post:
      tags: [Bus]
      summary: Queue deep bus scan
      description: |
        Accepts immediately (HTTP 202) and runs an extended reconcile in the
        background: full device rescan, extra settle (`bus.deep_settle_ms`),
        full enumeration, optional Give* probe transmits per device, then a
        broadcast Give Physical Address. Poll `GET /api/bus/state` until
        `scan_in_progress` is false. Vendor-specific skips via
        `bus.vendor_profiles` in `config.json`.
      operationId: postBusScan
      responses:
        '202':
          description: Deep scan queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '503':
          description: Steward queue full
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/topology:
    get:
      tags: [System]
      summary: CEC bus topology
      description: >
        Own logical addresses, active ports, and devices per port. Served from
        the cached steward snapshot (refresh is async); never blocks on the
        CEC connection.
      operationId: getTopology
      responses:
        '200':
          description: Topology retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────────── Power ───────────────────────────────
  /api/power/on:
    post:
      tags: [Power]
      summary: Power on TV
      description: Power on the TV (logical address 0).
      operationId: powerOn
      responses:
        '200':
          description: Power on command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/on/{address}:
    post:
      tags: [Power]
      summary: Power on device
      description: Power on a specific device by logical address.
      operationId: powerOnAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Power on command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/off:
    post:
      tags: [Power]
      summary: Power off TV
      description: Put the TV (logical address 0) in standby.
      operationId: powerOff
      responses:
        '200':
          description: Standby command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/off/{address}:
    post:
      tags: [Power]
      summary: Power off device
      description: Put a specific device in standby.
      operationId: powerOffAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Standby command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/status:
    get:
      tags: [Power]
      summary: Get TV power status
      description: Get power status of the TV (logical address 0).
      operationId: getPowerStatus
      responses:
        '200':
          description: Power status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                status: success
                message: Power status retrieved
                data:
                  address: 0
                  status: On
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/status/{address}:
    get:
      tags: [Power]
      summary: Get device power status
      description: Get power status of a specific device.
      operationId: getPowerStatusAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Power status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────────── Volume ──────────────────────────────
  /api/volume/up:
    post:
      tags: [Volume]
      summary: Volume up
      description: >
        Volume up via libcec (targets the TV's system audio path). Use the
        addressed variant to target another device (e.g. 5 = audio system).
      operationId: volumeUp
      responses:
        '200':
          description: Volume up sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/up/{address}:
    post:
      tags: [Volume]
      summary: Volume up to specific device
      operationId: volumeUpAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Volume up sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/down:
    post:
      tags: [Volume]
      summary: Volume down
      operationId: volumeDown
      responses:
        '200':
          description: Volume down sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/down/{address}:
    post:
      tags: [Volume]
      summary: Volume down to specific device
      operationId: volumeDownAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Volume down sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/mute:
    post:
      tags: [Volume]
      summary: Toggle mute
      operationId: volumeMute
      responses:
        '200':
          description: Mute toggle sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/mute/{address}:
    post:
      tags: [Volume]
      summary: Toggle mute on specific device
      operationId: volumeMuteAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Mute toggle sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────── Source / HDMI ───────────────────────────
  /api/source/active:
    get:
      tags: [Source]
      summary: Get current active source
      operationId: getSourceActive
      responses:
        '200':
          description: Active source retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/source/{address}:
    post:
      tags: [Source]
      summary: Switch to device by logical address
      operationId: postSourceAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Source switch sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/hdmi/{port}:
    post:
      tags: [Source]
      summary: Switch TV to HDMI port
      description: Switch the TV to HDMI port (1-15).
      operationId: postHdmiPort
      parameters:
        - name: port
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 15
      responses:
        '200':
          description: HDMI switch sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────────── Audio ───────────────────────────────
  /api/audio/status:
    get:
      tags: [System]
      summary: Audio status
      description: Volume level and mute state from the system audio device.
      operationId: getAudioStatus
      responses:
        '200':
          description: Audio status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ───────────────────────────── Navigation / Raw ─────────────────────────
  /api/key:
    post:
      tags: [Navigation]
      summary: Send key press
      description: |
        Send a remote key press by name (preferred) or raw keycode.
        **Keycode 0 is rejected** with an explicit error — keycode 0 is the
        Select code and must be sent as `{"key": "select"}` to avoid
        accidental selects from unset/zero-valued fields.
      operationId: postKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyRequest'
      responses:
        '200':
          description: Key press sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid address, unknown key name, or keycode 0 (use key:"select")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/command:
    post:
      tags: [Raw]
      summary: Send raw CEC command
      operationId: postCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
      responses:
        '200':
          description: Command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────── System routes ───────────────────────────
  /api/logs:
    get:
      tags: [System]
      summary: Recent CEC log messages
      operationId: getLogs
      responses:
        '200':
          description: Log buffer retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/update:
    post:
      tags: [System]
      summary: Trigger self-update
      description: |
        Downloads the latest release, verifies SHA256SUMS, swaps the binary
        (keeping `capi.bak` for rollback), and restarts the service. Never
        downgrades: the remote version must compare greater than the running
        semver. Returns `accepted=true` while the update runs in the
        background; progress lands in `/api/logs` and the SSE stream.
      operationId: postUpdate
      responses:
        '200':
          description: Update accepted or already up to date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                status: success
                message: update started
                data:
                  accepted: true
        '409':
          description: Update already in progress (single-flight)
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────── Settings ───────────────────────────────
  /api/settings/mqtt:
    get:
      tags: [Settings]
      summary: Get MQTT configuration and connection status
      operationId: getMqttSettings
      responses:
        '200':
          description: MQTT settings retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Settings]
      summary: Update MQTT settings
      description: |
        Persisted to `config.json`. The in-memory config only advances after
        the disk write succeeds (atomic update-and-persist under one lock).
      operationId: postMqttSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MQTTSettingsRequest'
      responses:
        '200':
          description: MQTT settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ───────────────────────────────── Dev routes ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional. Active only when the server runs with a token
        (`-token <secret>`). `X-Auth-Token`, `?key=`, and the `capi_token`
        cookie are accepted equivalently. When no token is configured this
        security requirement is a no-op.
  parameters:
    LogicalAddress:
      name: address
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 15
      description: CEC logical address (0-15; 15 is the broadcast address)
...
  responses:
...
    InternalError:
      description: CEC operation failed or server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: failed to open adapter